Split arch90.py into 7 focused rule modules
Replaces the monolithic arch90.py (18 rules, 6 concerns) with
descriptively-named modules: packaging, layers, config_rules, types,
errors, logging_rules, and ops. Test file split to match.

No rule logic changes — pure structural refactor.

Fixes #37
