+++ src/gaudi/packs/python/rules/__init__.py
@@ -15,7 +15,13 @@ from gaudi.packs.python.rules.pydantic import PYDANTIC_RULES
 from gaudi.packs.python.rules.pytest_rules import PYTEST_RULES
 from gaudi.packs.python.rules.drf import DRF_RULES
 from gaudi.packs.python.rules.smells import SMELL_RULES
-from gaudi.packs.python.rules.arch90 import ARCH90_RULES
+from gaudi.packs.python.rules.packaging import PACKAGING_RULES
+from gaudi.packs.python.rules.layers import LAYERS_RULES
+from gaudi.packs.python.rules.config_rules import CONFIG_RULES
+from gaudi.packs.python.rules.types import TYPES_RULES
+from gaudi.packs.python.rules.errors import ERRORS_RULES
+from gaudi.packs.python.rules.logging_rules import LOGGING_RULES
+from gaudi.packs.python.rules.ops import OPS_RULES
 from gaudi.packs.python.rules.stability import STABILITY_RULES
 from gaudi.packs.python.rules.services import SERVICE_RULES
 
@@ -33,7 +39,13 @@ ALL_RULES = (
     *PYTEST_RULES,
     *DRF_RULES,
     *SMELL_RULES,
-    *ARCH90_RULES,
+    *PACKAGING_RULES,
+    *LAYERS_RULES,
+    *CONFIG_RULES,
+    *TYPES_RULES,
+    *ERRORS_RULES,
+    *LOGGING_RULES,
+    *OPS_RULES,
     *STABILITY_RULES,
     *SERVICE_RULES,
 )
